Facebook create channel
I have a small list of elite programming ... choose a name lightly, there may be a few in this list that just think I'm elite and you do not attract everyone at all. Also, the best way to find smart people to track is to look at your respected programmers and follow them. And if you do not mind listening to a little grunt, you can always follow me